minor edits and design fix

+++ manager/src/Model/Work/UseCase/Members/Group/Remove/Handler.php
@@ -7,6 +7,7 @@ namespace App\Model\Work\UseCase\Members\Group\Remove;
 use App\Model\Flusher;
 use App\Model\Work\Entity\Members\Group\GroupRepository;
 use App\Model\Work\Entity\Members\Group\Id;
+use App\Model\Work\Entity\Members\Member\MemberRepository;
 
 /**
  * Class Handler
@@ -22,16 +23,22 @@ class Handler
      * @var Flusher
      */
     private $flusher;
+    /**
+     * @var MemberRepository
+     */
+    private $memberRepository;
 
     /**
      * Handler constructor.
      * @param GroupRepository $groupRepository
+     * @param MemberRepository $memberRepository
      * @param Flusher $flusher
      */
-    public function __construct(GroupRepository $groupRepository, Flusher $flusher)
+    public function __construct(GroupRepository $groupRepository, MemberRepository $memberRepository, Flusher $flusher)
     {
         $this->groupRepository = $groupRepository;
         $this->flusher = $flusher;
+        $this->memberRepository = $memberRepository;
     }
 
     /**
@@ -40,6 +47,10 @@ class Handler
     public function handle(Command $command)
     {
         $group = $this->groupRepository->get(new Id($command->id));
+
+        if ($this->memberRepository->hasByGroup($group->getId())) {
+            throw new \DomainException('Group is not empty');
+        }
 
         $this->groupRepository->remove($group);
